+++ src/main.rs
@@ -1,9 +1,11 @@
 mod app;
-
-use app::utils::get_args;
+use brr::get_args;
 
 fn main() {
     let args = get_args();
-    let mut app = app::Session::new(args);
+    let action = app::bash_driver::get_action(&args);
+    let arguments = app::bash_driver::get_argument(&mut args);
+
+    let mut app = app::Session::run(action);
     app.run();
 }
